requirements.txt
-i https://pypi.org/simple
aiofiles==24.1.0; python_version >= '3.8'
alembic==1.15.1; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
apscheduler==3.11.0; python_version >= '3.8'
asyncpg==0.30.0; python_full_version >= '3.8.0'
certifi==2025.1.31; python_version >= '3.6'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
dnspython==2.7.0; python_version >= '3.9'
ecdsa==0.19.1; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
email-validator==2.2.0; python_version >= '3.8'
fastapi[standard]==0.115.11; python_version >= '3.8'
fastapi-cli[standard]==0.0.7; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
mako==1.3.9; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pydantic-settings==2.8.1; python_version >= '3.8'
pygments==2.19.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-jose==3.4.0
python-multipart==0.0.20; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
rich==13.9.4; python_full_version >= '3.8.0'
rich-toolkit==0.13.2; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.39; python_version >= '3.7'
starlette==0.46.1; python_version >= '3.9'
typer==0.15.2; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
tzlocal==5.3.1; python_version >= '3.9'
uvicorn[standard]==0.34.0; python_version >= '3.9'
watchfiles==1.0.4; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
